Description of Projects
1) House Price Prediction
This Dataset used for the Prediction of house price. From the dataset predict the house price using various machine learning algorithm. So, from various algorithm select the best suitable algorithm which gives the best score. To increase the accuracy done the cross validation and hyper tuning parameter.

2) Product Matching
In this dataset merge the two different dataset. The dataset are e-commerce plateform flipkart and amazon used. So, these dataset merge for to get the difference in the price of product which buy from these platform.

3) Children States
This Dataset used to find the information regarding children of different age group. Used the different visualization library to get insight among children. In this get insights from different charts.